TIP-885: move forms
