+++ UPGRADE.md
@@ -55,6 +55,7 @@ Several classes and services have been moved or renamed. The following commands 
 
 ```bash
 
+find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\LocalizationBundle\\Form\\Type\\LocaleType/Akeneo\\Platform\\Bundle\\UIBundle\\Form\\Type\\LocaleType/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\LocalizationBundle\\Twig\\AttributeExtension/Akeneo\\Platform\\Bundle\\UIBundle\\Twig\\AttributeExtension/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\LocalizationBundle\\Twig\\LocaleExtension/Akeneo\\Platform\\Bundle\\UIBundle\\Twig\\LocaleExtension/g'
 find ./src/ -type f -print0 | xargs -0 sed -i 's/Pim\\Bundle\\ReferenceDataBundle\\DataGrid\\Extension\\Sorter\\ReferenceDataSorter/Oro\\Bundle\\PimDataGridBundle\\Extension\\Sorter\\Produc\\ReferenceDataSorter/g'
